Move PitchProjector into services
