+++ include/fullscore/services/pitch_projector.h
